LF-4163 Move field names and types to types.ts file; nest containers
