+++ packages/webapp/src/containers/Animals/AddAnimals/AddAnimalBasics/index.tsx
@@ -22,14 +22,14 @@ import {
   useGetDefaultAnimalBreedsQuery,
   useGetCustomAnimalBreedsQuery,
   useGetAnimalSexesQuery,
-} from '../../../store/api/apiSlice';
-import AddAnimalsFormCard from '../../../components/Animals/AddAnimalsFormCard/AddAnimalsFormCard';
-import MoreAnimalsCard from '../../../components/Animals/AddAnimalsForm/MoreAnimalsCard';
-import { FIELD_NAMES } from '../../../components/Animals/AddAnimalsFormCard/AddAnimalsFormCard';
-import { AddAnimalsFormFields } from '../AddAnimals/types';
-import { generateUniqueAnimalId } from '../../../util/animal';
-import { ANIMAL_ID_PREFIX } from '../types';
-import { STEPS } from '../AddAnimals';
+} from '../../../../store/api/apiSlice';
+import AddAnimalsFormCard from '../../../../components/Animals/AddAnimalsFormCard/AddAnimalsFormCard';
+import MoreAnimalsCard from '../../../../components/Animals/AddAnimalsForm/MoreAnimalsCard';
+import { ANIMAL_BASICS_FIELD_NAMES as FIELD_NAMES } from '../types';
+import { AddAnimalsFormFields } from '../types';
+import { generateUniqueAnimalId } from '../../../../util/animal';
+import { ANIMAL_ID_PREFIX } from '../../types';
+import { STEPS } from '..';
 
 export const animalBasicsDefaultValues = {
   [FIELD_NAMES.TYPE]: undefined,
