+++ packages/webapp/src/containers/Animals/AddAnimals/AddAnimalBasics/styles.module.scss
@@ -13,7 +13,7 @@
  *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
  */
 
-@import '../../../assets/mixin.scss';
+@import '../../../../assets/mixin.scss';
 
 .cardContainer {
   padding: 16px;
