+++ packages/webapp/src/containers/Animals/AddAnimals/AddAnimalDetails/index.tsx
